Add number formatting, fix tests, code cleanup

+++ domain/src/main/java/pl/mwaszczuk/domain/model/Currency.kt
@@ -1,8 +1,9 @@
 package pl.mwaszczuk.domain.model
 
 import androidx.annotation.StringRes
+import java.math.BigDecimal
 
 data class Currency(
     @StringRes val currencyStringRes: Int,
-    val amount: Double
+    val amount: BigDecimal
 )
